reworked 'gallery' ui to create centered grids
fell back to using flexbox with media queries, instead of the css grid we're using on carmine. grid is generally better, but does not allow you to center-justify the bottom row, and left-justified rows look weird on our homepage.

we should be able to get away with using media queries on grids for now b/c all grids are full-width. if our layouts get more complex, we could switch to a JS container-query polyfill.